hackathon-project/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
python-multipart==0.0.9
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# torch==2.3.0          # Uncomment for PyTorch
# tensorflow==2.16.1    # Uncomment for TensorFlow
# transformers==4.41.1  # Uncomment for HuggingFace
